Add content section
Content section currently has 3 main sections, the left for a picture of
me, the middle for some info about me, and the right will list
tehcnologies I know. Font awesome remains to be completely set up for
this purpose, and the footer must be created. Also in the future is
creating the projects page, which I may create an API for just to
simplify the display of new ones in the future.