Please help. Add flexbox project layout and add responsive breakpoints.
Adds a url field to the project json file. The images
will be kept in the static directory, and the urls added along with the
descriptions. Will likely add image location field for json to make it
convenient to add them later. Added flexbox based system for creating
the project list. Still need to tweak column counts and things later,
but what matters is that I have a system I'm pretty okay with.